0.2.0.9000 -- Major revisions to API
* New API for setting metadata. Setting metadata for any object is now
accomplished with `add_metadata(obj, ...)` or via a new assignment
operator (`metadata(obj) <- list(...)`). The functions of the old
`add_metadata` have been replaced with `add_column_metadata`.

* An additional type conversion step has been added to `read_csvy`,
since `fread` will not actually respect `colClasses` if it thinks doing
so will result in data loss. Mismatches in `fread`'s guess and the
specified classes will result in a warning that can be suppressed with
`verbose = FALSE`.

* Lots of new tests to check new functionality. All tests are passing.